requirements.txt
streamlit
tensorflow
numpy
pillow
pandas
gdown